# Backend/swagger.yaml
openapi: 3.0.0
info:
  title: User Management API
  description: API for managing user accounts and events
  version: 1.0.0
servers:
  - url: /api/v1
    description: Main API server
security:
  - bearerAuth: []
paths:
...
  /events:
    post:
      tags:
        - Events
      summary: Create a new event
      description: Creates an event with a title, organizer, and other fields.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Invalid input or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Events
      summary: Get all events
      description: Retrieve a paginated and filterable list of events
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - name: title
          in: query
          schema:
            type: string
          description: Filter by event title
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: eventType
          in: query
          schema:
            type: string
          description: Filter by event type
        - name: status
          in: query
          schema:
            type: string
          description: Filter by status
        - name: organizer
          in: query
          schema:
            type: string
          description: Filter by organizer ID
      responses:
        '200':
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /events/{id}:
    get:
      tags:
        - Events
      summary: Get event by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      responses:
        '200':
          description: Event retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Events
      summary: Update event by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Events
      summary: Delete event by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/eventId'
      responses:
        '204':
          description: Event deleted successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Registration APIs
  /registrations:
    post:
      tags:
        - Registrations
      summary: Create a new registration
      description: Registers a user for an event.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user
                - event
              properties:
                user:
                  type: string
                  description: ID of the user (pointer to _User)
                  example: "abc123userId"
                event:
                  type: string
                  description: ID of the event (pointer to Event)
                  example: "xyz789eventId"
                status:
                  type: string
                  enum: [pending, confirmed, cancelled]
                  description: Registration status (optional, defaults to pending)
                  example: "pending"
                registeredAt:
                  type: string
                  format: date-time
                  description: Date and time of registration (optional)
                  example: "2025-07-15T10:00:00Z"
      responses:
        '201':
          description: Registration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          description: Validation error or missing required fields
        '409':
          description: Duplicate registration

    get:
      tags:
        - Registrations
      summary: Get a list of registrations
      description: Retrieve a paginated list of registrations with optional filters and sorting.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/sortBy'
        - $ref: '#/components/parameters/sortOrder'
        - name: userId
          in: query
          schema:
            type: string
          description: Filter by user ID
        - name: eventId
          in: query
          schema:
            type: string
          description: Filter by event ID
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled]
          description: Filter by registration status
        - name: fromDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by registeredAt >= fromDate
        - name: toDate
          in: query
          schema:
            type: string
            format: date-time
          description: Filter by registeredAt <= toDate
      responses:
        '200':
          description: A list of registrations matching the filters
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/RegistrationResponse'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      total:
                        type: integer
                        example: 42
        '400':
          description: Bad request
  /registrations/{id}:
    get:
      tags:
        - Registrations
      summary: Get a registration by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Registration ID
      responses:
        '200':
          description: Registration fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '404':
          description: Registration not found

    put:
      tags:
        - Registrations
      summary: Update a registration by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Registration ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: string
                  description: Updated user ID
                  example: "abc123userId"
                event:
                  type: string
                  description: Updated event ID
                  example: "xyz789eventId"
                status:
                  type: string
                  enum: [pending, confirmed, cancelled]
                  description: Updated registration status
                  example: "confirmed"
                registeredAt:
                  type: string
                  format: date-time
                  description: Updated registration timestamp
                  example: "2025-07-15T10:00:00Z"
      responses:
        '200':
          description: Registration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistrationResponse'
        '400':
          description: Validation error
        '404':
          description: Registration not found

    delete:
      tags:
        - Registrations
      summary: Delete a registration by ID
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Registration ID
      responses:
        '204':
          description: Registration deleted successfully
        '404':
          description: Registration not found
  # Event statistc API's
  /statistics/events/count:
    get:
      tags:
        - Statistics
      summary: Get total number of events
      description: Returns the total count of events in the system.
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalEvents:
                    type: integer
                    example: 42
        '500':
          description: Internal Server Error
  /statistics/events/status:
    get:
      tags:
        - Statistics
      summary: Get count of events by status
      description: Returns event counts grouped by status (e.g., active, completed).
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                example:
                  active: 10
                  canceled: 5
                  completed: 15
                  scheduled: 12
        '500':
          description: Internal Server Error
  /statistics/registrations/trends:
    get:
      tags:
        - Statistics
      summary: Get registration trends
      description: Returns number of registrations per day between given dates.
      parameters:
        - name: fromDate
          in: query
          description: Start date in ISO format (e.g., 2024-01-01)
          required: false
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: End date in ISO format (e.g., 2024-12-31)
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      example: 2024-05-01
                    registrations:
                      type: integer
                      example: 12
        '500':
          description: Internal Server Error
  /statistics/events/popular:
    get:
      tags:
        - Statistics
      summary: Get most popular events
      description: Returns events with the highest number of registrations.
      parameters:
        - name: limit
          in: query
          description: Number of top events to return (default is 5)
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    eventId:
                      type: string
                      example: "abc123"
                    title:
                      type: string
                      example: "Summer Tech Meetup"
                    registrations:
                      type: integer
                      example: 89
        '500':
          description: Internal Server Error
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
components:
  parameters:
...
    eventId:
      name: id
      in: path
      required: true
      description: ID of the event
      schema:
        type: string
        example: e1234567890
    page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        default: 1
        minimum: 1
    limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    sortBy:
      name: sortBy
      in: query
      description: Sort by field
      schema:
        type: string
        enum: [createdAt, updatedAt, username, email]
        default: createdAt
    sortOrder:
      name: sortOrder
      in: query
      description: Sort direction
      schema:
        type: string
        enum: [asc, desc]
        default: desc
  schemas:
...
    EventCreate:
      type: object
      required:
        - title
        - description
        - location
        - startDate
        - endDate
        - maxParticipants
        - eventType
        - status
        - organizer
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        maxParticipants:
          type: integer
        eventType:
          type: string
        status:
          type: string
        organizer:
          type: string
          description: ID of the user organizing the event
    Event:
      type: object
      properties:
        objectId:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        maxParticipants:
          type: integer
        eventType:
          type: string
        status:
          type: string
        organizer:
          type: object
          properties:
            objectId:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RegistrationResponse:
      type: object
      properties:
        objectId:
          type: string
          example: "abc123xyz"
        user:
          type: object
          description: Pointer to user object
          example:
            objectId: "abc123userId"
            className: "_User"
        event:
          type: object
          description: Pointer to event object
          example:
            objectId: "xyz789eventId"
            className: "Event"
        status:
          type: string
          example: pending
        registeredAt:
          type: string
          format: date-time
          example: "2025-07-15T10:00:00Z"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Error message describing the issue
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT